feat: add create anonymous context to agent
