+++ lib/egg.js
@@ -3,6 +3,7 @@
 const path = require('path');
 const fs = require('fs');
 const ms = require('ms');
+const http = require('http');
 const EggCore = require('egg-core').EggCore;
 const cluster = require('cluster-client');
 const extend = require('extend2');
@@ -456,6 +457,75 @@ class EggApplication extends EggCore {
       return realClient;
     };
   }
+
+  /**
+   * Create an anonymous context, the context isn't request level, so the request is mocked.
+   * then you can use context level API like `ctx.service`
+   * @member {String} EggApplication#createAnonymousContext
+   * @param {Request} req - if you want to mock request like querystring, you can pass an object to this function.
+   * @return {Context} context
+   */
+  createAnonymousContext(req) {
+    const request = {
+      headers: {
+        'x-forwarded-for': '127.0.0.1',
+      },
+      query: {},
+      querystring: '',
+      host: '127.0.0.1',
+      hostname: '127.0.0.1',
+      protocol: 'http',
+      secure: 'false',
+      method: 'GET',
+      url: '/',
+      path: '/',
+      socket: {
+        remoteAddress: '127.0.0.1',
+        remotePort: 7001,
+      },
+    };
+    if (req) {
+      for (const key in req) {
+        if (key === 'headers' || key === 'query' || key === 'socket') {
+          Object.assign(request[key], req[key]);
+        } else {
+          request[key] = req[key];
+        }
+      }
+    }
+    const response = new http.ServerResponse(request);
+    return this.createContext(request, response);
+  }
+
+  /**
+   * Create egg context
+   * @method EggApplication#createContext
+   * @param  {Req} req - node native Request object
+   * @param  {Res} res - node native Response object
+   * @return {Context} context object
+   */
+  createContext(req, res) {
+    const app = this;
+    const context = Object.create(app.context);
+    const request = context.request = Object.create(app.request);
+    const response = context.response = Object.create(app.response);
+    context.app = request.app = response.app = app;
+    context.req = request.req = response.req = req;
+    context.res = request.res = response.res = res;
+    request.ctx = response.ctx = context;
+    request.response = response;
+    response.request = request;
+    context.onerror = context.onerror.bind(context);
+    context.originalUrl = request.originalUrl = req.url;
+
+    /**
+     * Request start time
+     * @member {Number} Context#starttime
+     */
+    context.starttime = Date.now();
+    return context;
+  }
+
 }
 
 module.exports = EggApplication;
